testing

+++ src/vsr/models/VRT/modules/stochastic_depth.py
@@ -7,10 +7,10 @@ def drop_path(x, drop_prob: float = 0., training: bool = False):
         return x
     keep_prob = 1 - drop_prob
     shape = (x.shape[0],) + (1,) * (x.ndim - 1)  # work with diff dim tensors, not just 2D ConvNets
-    random_tensor = keep_prob + torch.rand(shape).type_as(x)
+    random_tensor = keep_prob + torch.rand(shape)
     random_tensor.floor_()  # binarize
-    output = x.div(keep_prob) * random_tensor
-    return output
+    output = x.div(keep_prob) * random_tensor.type_as(x)
+    return output.type_as(x)
 
 class DropPath(nn.Module):
     """Drop paths (Stochastic Depth) per sample  (when applied in main path of residual blocks)"""
